add tasks and htmltasks of literary movie,and add test cases of them.Add the subclass LiteraryAmount extends Amount and implements the methods getPriceCode and getFinalAmount.Change the return type of method getTotalFrequentRenterPoints from int to String,and change the logic,when the score is end with ". 0",remove ". 0" and return String.valueOf(result),else return String.valueOf(result).

+++ src/main/java/rentalstore/Rental.java
@@ -13,7 +13,7 @@ public class Rental {
         return movie;
     }
 
-    int getFrequentRentalPoints() {
+    double getFrequentRentalPoints() {
         return movie.getFrequentRentalPoints(dayRented);
     }
     double getFinalAmount(){
